Added a more complex test case
